+++ go/integTest/readfrom_strategy_test.go
@@ -10,9 +10,9 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/valkey-io/valkey-glide/go/api"
-	"github.com/valkey-io/valkey-glide/go/api/config"
-	"github.com/valkey-io/valkey-glide/go/api/options"
+	"github.com/valkey-io/valkey-glide/go/v2/config"
+	"github.com/valkey-io/valkey-glide/go/v2/constants"
+	"github.com/valkey-io/valkey-glide/go/v2/options"
 )
 
 func (suite *GlideTestSuite) TestRoutingWithAzAffinityStrategyTo1Replica() {
@@ -41,7 +41,7 @@ func (suite *GlideTestSuite) TestRoutingWithAzAffinityStrategyTo1Replica() {
 
 	clientForTestingAz := suite.clusterClient(suite.defaultClusterClientConfig().
 		WithRequestTimeout(2 * time.Second).
-		WithReadFrom(api.AzAffinity).
+		WithReadFrom(config.AzAffinity).
 		WithClientAZ(az))
 
 	for i := 0; i < GET_CALLS; i++ {
@@ -51,7 +51,7 @@ func (suite *GlideTestSuite) TestRoutingWithAzAffinityStrategyTo1Replica() {
 
 	infoResult, err := clientForTestingAz.InfoWithOptions(context.Background(),
 		options.ClusterInfoOptions{
-			InfoOptions: &options.InfoOptions{Sections: []options.Section{options.Server, options.Commandstats}},
+			InfoOptions: &options.InfoOptions{Sections: []constants.Section{constants.Server, constants.Commandstats}},
 			RouteOption: &options.RouteOption{Route: config.AllNodes},
 		},
 	)
@@ -95,7 +95,7 @@ func (suite *GlideTestSuite) TestRoutingBySlotToReplicaWithAzAffinityStrategyToA
 	clusterInfo, err := clientForConfigSet.InfoWithOptions(context.Background(),
 		options.ClusterInfoOptions{
 			RouteOption: &options.RouteOption{Route: config.NewSlotKeyRoute(config.SlotTypePrimary, "key")},
-			InfoOptions: &options.InfoOptions{Sections: []options.Section{options.Replication}},
+			InfoOptions: &options.InfoOptions{Sections: []constants.Section{constants.Replication}},
 		})
 	assert.NoError(suite.T(), err)
 	nReplicas := 0
@@ -119,7 +119,7 @@ func (suite *GlideTestSuite) TestRoutingBySlotToReplicaWithAzAffinityStrategyToA
 	// Creating Client with AZ configuration for testing
 	clientForTestingAz := suite.clusterClient(suite.defaultClusterClientConfig().
 		WithRequestTimeout(2 * time.Second).
-		WithReadFrom(api.AzAffinity).
+		WithReadFrom(config.AzAffinity).
 		WithClientAZ(az))
 
 	azGetResult, err := clientForTestingAz.ConfigGetWithOptions(context.Background(),
@@ -137,7 +137,7 @@ func (suite *GlideTestSuite) TestRoutingBySlotToReplicaWithAzAffinityStrategyToA
 
 	infoResult, err := clientForTestingAz.InfoWithOptions(context.Background(),
 		options.ClusterInfoOptions{
-			InfoOptions: &options.InfoOptions{Sections: []options.Section{options.All}},
+			InfoOptions: &options.InfoOptions{Sections: []constants.Section{constants.All}},
 			RouteOption: &options.RouteOption{Route: config.AllNodes},
 		},
 	)
@@ -162,11 +162,11 @@ func (suite *GlideTestSuite) TestAzAffinityNonExistingAz() {
 	const nReplicaCalls = 1
 	getCmdStat := fmt.Sprintf("cmdstat_get:calls=%d", nReplicaCalls)
 
-	clientForTestingAz := suite.clusterClient(api.NewGlideClusterClientConfiguration().
+	clientForTestingAz := suite.clusterClient(config.NewClusterClientConfiguration().
 		WithAddress(&suite.clusterHosts[0]).
 		WithUseTLS(suite.tls).
 		WithRequestTimeout(2 * time.Second).
-		WithReadFrom(api.AzAffinity).
+		WithReadFrom(config.AzAffinity).
 		WithClientAZ("non-existing-az"))
 
 	// Reset stats
@@ -182,7 +182,7 @@ func (suite *GlideTestSuite) TestAzAffinityNonExistingAz() {
 
 	infoResult, err := clientForTestingAz.InfoWithOptions(context.Background(),
 		options.ClusterInfoOptions{
-			InfoOptions: &options.InfoOptions{Sections: []options.Section{options.Commandstats}},
+			InfoOptions: &options.InfoOptions{Sections: []constants.Section{constants.Commandstats}},
 			RouteOption: &options.RouteOption{Route: config.AllNodes},
 		},
 	)
@@ -209,7 +209,7 @@ func (suite *GlideTestSuite) TestAzAffinityReplicasAndPrimaryRoutesToPrimary() {
 	getCmdStat := fmt.Sprintf("cmdstat_get:calls=%d", nGetCalls)
 
 	// Create client for setting the configs
-	clientForConfigSet := suite.clusterClient(api.NewGlideClusterClientConfiguration().
+	clientForConfigSet := suite.clusterClient(config.NewClusterClientConfiguration().
 		WithAddress(&suite.clusterHosts[0]).
 		WithUseTLS(suite.tls).
 		WithRequestTimeout(2 * time.Second))
@@ -242,11 +242,11 @@ func (suite *GlideTestSuite) TestAzAffinityReplicasAndPrimaryRoutesToPrimary() {
 	clientForConfigSet.Close()
 
 	// Create test client with AZ_AFFINITY_REPLICAS_AND_PRIMARY configuration
-	clientForTestingAz := suite.clusterClient(api.NewGlideClusterClientConfiguration().
+	clientForTestingAz := suite.clusterClient(config.NewClusterClientConfiguration().
 		WithAddress(&suite.clusterHosts[0]).
 		WithUseTLS(suite.tls).
 		WithRequestTimeout(2 * time.Second).
-		WithReadFrom(api.AzAffinityReplicaAndPrimary).
+		WithReadFrom(config.AzAffinityReplicaAndPrimary).
 		WithClientAZ(az))
 
 	// Execute GET commands
@@ -257,7 +257,7 @@ func (suite *GlideTestSuite) TestAzAffinityReplicasAndPrimaryRoutesToPrimary() {
 
 	infoResult, err := clientForTestingAz.InfoWithOptions(context.Background(),
 		options.ClusterInfoOptions{
-			InfoOptions: &options.InfoOptions{Sections: []options.Section{options.All}},
+			InfoOptions: &options.InfoOptions{Sections: []constants.Section{constants.All}},
 			RouteOption: &options.RouteOption{Route: config.AllNodes},
 		},
 	)